Fix cyclic build dependency for tests
