+++ gossip/store_llr_block.go
@@ -1,10 +1,8 @@
 package gossip
 
 import (
-	"github.com/Fantom-foundation/go-opera/inter"
 	"github.com/Fantom-foundation/go-opera/inter/ibr"
 	"github.com/Fantom-foundation/go-opera/inter/ier"
-	"github.com/Fantom-foundation/lachesis-base/hash"
 	"github.com/Fantom-foundation/lachesis-base/inter/idx"
 	"github.com/ethereum/go-ethereum/core/types"
 )
@@ -19,7 +17,7 @@ func (s *Store) GetFullBlockRecord(n idx.Block) *ibr.LlrFullBlockRecord {
 	if receipts == nil {
 		receipts = []*types.ReceiptForStorage{}
 	}
-	return FullBlockRecordFor(block, txs, receipts)
+	return ibr.FullBlockRecordFor(block, txs, receipts)
 }
 
 func (s *Store) GetFullEpochRecord(epoch idx.Epoch) *ier.LlrFullEpochRecord {
@@ -30,26 +28,5 @@ func (s *Store) GetFullEpochRecord(epoch idx.Epoch) *ier.LlrFullEpochRecord {
 	return &ier.LlrFullEpochRecord{
 		BlockState: *hbs,
 		EpochState: *hes,
-	}
-}
-
-// FullBlockRecordFor returns the full block record used in Genesis processing
-// for the given block, list of transactions, and list of transaction receipts.
-func FullBlockRecordFor(block *inter.Block, txs types.Transactions,
-	rawReceipts []*types.ReceiptForStorage) *ibr.LlrFullBlockRecord {
-	return &ibr.LlrFullBlockRecord{
-		BlockHash:  hash.Hash(block.Hash()),
-		ParentHash: hash.Hash(block.ParentHash),
-		StateRoot:  hash.Hash(block.StateRoot),
-		Time:       block.Time,
-		Duration:   block.Duration,
-		Difficulty: block.Difficulty,
-		GasLimit:   block.GasLimit,
-		GasUsed:    block.GasUsed,
-		BaseFee:    block.BaseFee,
-		PrevRandao: hash.Hash(block.PrevRandao),
-		Epoch:      block.Epoch,
-		Txs:        txs,
-		Receipts:   rawReceipts,
 	}
 }
