+++ integration/makegenesis/genesis.go
@@ -5,13 +5,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/Fantom-foundation/go-opera/gossip"
-	"github.com/Fantom-foundation/go-opera/inter"
-	"github.com/ethereum/go-ethereum/core/tracing"
 	"io"
 	"math/big"
 	"os"
 	"path/filepath"
+
+	"github.com/Fantom-foundation/go-opera/inter"
+	"github.com/ethereum/go-ethereum/core/tracing"
 
 	"github.com/Fantom-foundation/lachesis-base/hash"
 	"github.com/ethereum/go-ethereum/common"
@@ -220,7 +220,7 @@ func (b *GenesisBuilder) ExecuteGenesisTxs(blockProc BlockProc, genesisTxs types
 		blockBuilder.AddTransaction(transaction, receipts[txIndex])
 	}
 
-	llrBlock := gossip.FullBlockRecordFor(blockBuilder.Build(), evmBlock.Transactions, receiptsStorage)
+	llrBlock := ibr.FullBlockRecordFor(blockBuilder.Build(), evmBlock.Transactions, receiptsStorage)
 	b.blocks = append(b.blocks, ibr.LlrIdxFullBlockRecord{
 		LlrFullBlockRecord: *llrBlock,
 		Idx:                blockCtx.Idx,
