+++ inter/ibr/inter_block_records.go
@@ -1,11 +1,12 @@
 package ibr
 
 import (
+	"math/big"
+
 	"github.com/Fantom-foundation/go-opera/inter"
 	"github.com/Fantom-foundation/lachesis-base/hash"
 	"github.com/Fantom-foundation/lachesis-base/inter/idx"
 	"github.com/ethereum/go-ethereum/core/types"
-	"math/big"
 )
 
 type LlrFullBlockRecord struct {
@@ -27,4 +28,25 @@ type LlrFullBlockRecord struct {
 type LlrIdxFullBlockRecord struct {
 	LlrFullBlockRecord
 	Idx idx.Block
+}
+
+// FullBlockRecordFor returns the full block record used in Genesis processing
+// for the given block, list of transactions, and list of transaction receipts.
+func FullBlockRecordFor(block *inter.Block, txs types.Transactions,
+	rawReceipts []*types.ReceiptForStorage) *LlrFullBlockRecord {
+	return &LlrFullBlockRecord{
+		BlockHash:  hash.Hash(block.Hash()),
+		ParentHash: hash.Hash(block.ParentHash),
+		StateRoot:  hash.Hash(block.StateRoot),
+		Time:       block.Time,
+		Duration:   block.Duration,
+		Difficulty: block.Difficulty,
+		GasLimit:   block.GasLimit,
+		GasUsed:    block.GasUsed,
+		BaseFee:    block.BaseFee,
+		PrevRandao: hash.Hash(block.PrevRandao),
+		Epoch:      block.Epoch,
+		Txs:        txs,
+		Receipts:   rawReceipts,
+	}
 }
